+++ R/assay5.R
@@ -1518,7 +1518,7 @@ SetAssayData.StdAssay <- function(
     )
     layer <- slot
   }
-  LayerData(object = object, layer = slot) <- new.data
+  LayerData(object = object, layer = layer) <- new.data
   return(object)
 }
 
@@ -2733,7 +2733,7 @@ setAs(
     # browser()
     # Add the expression matrices
     for (i in c('counts', 'data', 'scale.data')) {
-      adata <- GetAssayData(object = from, slot = i)
+      adata <- GetAssayData(object = from, layer = i)
       if (IsMatrixEmpty(x = adata)) {
         next
       }